pyproject.toml
[project]
name = "serializd-importer"
version = "0.2.0"
dependencies = [
    "tmdbsimple",
    "python-dotenv",
]

[project.scripts]
serializd-importer = "serializd_importer.cli:main"

[tool.setuptools]
packages = ["serializd_importer"]
package-dir = {"" = "src"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Development: editable install of serializd-py
# Install with: pip install -e ../serializd-py
